+++ packages/anvl/src/id/identified.ts
@@ -1,6 +1,5 @@
+import type { Json } from "~/packages/atom.io/json/src"
 import type { Refinement } from "~/packages/rel8/types/src"
-
-import type { Json } from "../json"
 
 export type Identified = { id: string }
 export type Branded<TypeName extends string> = { type: TypeName }
