+++ packages/anvl/src/join/index.ts
@@ -1,5 +1,6 @@
+import type { Json, JsonInterface } from "atom.io/json"
+
 import type { Identified } from "../id"
-import type { Json, JsonInterface } from "../json"
 import type { NullSafeRest, NullSafeUnion } from "../nullish"
 import type { Refinement } from "../refinement"
 import type { IsRelationDataOptions, RelationData } from "./core-relation-data"
