--- packages/atom.io/__unstable__/web-effects/src/index.ts
@@ -1,1 +1,0 @@
-export * from "./storage"
