+++ packages/hamr/react-json-editor/src/editors-by-type/utilities/array-elements.ts
@@ -1,7 +1,7 @@
+import type { Json } from "atom.io/json"
 import type { SetterOrUpdater } from "recoil"
 
 import { become } from "~/packages/anvl/src/function"
-import type { Json } from "~/packages/anvl/src/json"
 
 export const makeElementSetters = <T extends Json.Array>(
 	data: T,
