+++ packages/hamr/react-json-editor/src/editors-by-type/utilities/object-properties.ts
@@ -2,10 +2,11 @@ import type { MutableRefObject } from "react"
 import type { SetterOrUpdater } from "recoil"
 
 import { become } from "~/packages/anvl/src/function"
-import type { Json, JsonTypeName } from "~/packages/anvl/src/json"
+import type { JsonTypeName } from "~/packages/anvl/src/json"
 import { JSON_DEFAULTS } from "~/packages/anvl/src/json"
 import { mapObject } from "~/packages/anvl/src/object"
 import { castToJson } from "~/packages/anvl/src/refinement/smart-cast-json"
+import type { Json } from "~/packages/atom.io/json/src"
 
 export const makePropertySetters = <T extends Json.Object>(
 	data: T,
